cleanup implementation
